Refactor spinner
